+++ scripts/Models/Spell.js
@@ -102,6 +102,16 @@ export default class Spell {
     {id:'sacrement-otherworldly', name: 'Otherworldly Sacrement', category: 'tools', bonus: 3, desc: '<p>A sacrement from somewhere other than the material realm.</p>'},
   ];
 
+  static diceSizes = [
+    { id: "none", name: "None" },
+    { id: "d4", name: "d4" },
+    { id: "d6", name: "d6" },
+    { id: "d8", name: "d8 (Bashing)" },
+    { id: "d10", name: "d10 (Lethal)" },
+    { id: "d12", name: "d12 (Aggravated)" },
+    { id: "d20", name: "d20" },
+  ];
+
   static rankedPractices(highestArcanum, showAll)
   {
     return this.practices.reduce((acc, p) => {